# Use an efficient queue implementation