--- a/index.js
+++ b/index.js
@@ -3,26 +3,26 @@
 module.exports = function (PromiseArgument) {
   var Promise;
   function throat(size, fn) {
-    var queue = []
+    var queue = [];
     function run(fn, self, args) {
       if (size) {
-        size--
+        size--;
         var result = new Promise(function (resolve) {
-          resolve(fn.apply(self, args))
-        })
-        result.then(release, release)
-        return result
+          resolve(fn.apply(self, args));
+        });
+        result.then(release, release);
+        return result;
       } else {
         return new Promise(function (resolve) {
-          queue.push(new Delayed(resolve, fn, self, args))
-        })
+          queue.push(new Delayed(resolve, fn, self, args));
+        });
       }
     }
     function release() {
-      size++
+      size++;
       if (queue.length) {
-        var next = queue.shift()
-        next.resolve(run(next.fn, next.self, next.args))
+        var next = queue.shift();
+        next.resolve(run(next.fn, next.self, next.args));
       }
     }
     if (typeof size === 'function') {
@@ -83,8 +83,8 @@ if (typeof Promise === 'function') {
 }
 
 function Delayed(resolve, fn, self, args) {
-  this.resolve = resolve
-  this.fn = fn
-  this.self = self || null
-  this.args = args
+  this.resolve = resolve;
+  this.fn = fn;
+  this.self = self || null;
+  this.args = args;
 }