--- a/package.json
+++ b/package.json
@@ -11,7 +11,9 @@
     "parallelism",
     "limit"
   ],
-  "dependencies": {},
+  "dependencies": {
+    "double-ended-queue": "^2.1.0-0"
+  },
   "devDependencies": {
     "coveralls": "^2.11.2",
     "istanbul": "^0.4.5",