I'm dealing with 100,000 to 1 million promises. Don't blame me, blame async/await. I noticed poor runtime performance of my code and thought it was because of what I was doing in the promises but after profiling performance about 95% of the time in my program was spent removing the first item from large arrays. This makes sense: in v8, this operation is linear but appears fast for small arrays because of some trickery.

This diff:
* Introduces a performance benchmark that works in node 7+ (requires async/await). I don't think it makes much sense to support older versions of node for this one.
* Swaps out the array queue with [`double-ended-queue`](https://yarnpkg.com/en/package/double-ended-queue).

Before:
```
[cpojer ~/P/throat (master|ahead)]$ node perf.js
10 promises: 1.709ms
100 promises: 5.487ms
1000 promises: 17.602ms
10000 promises: 218.416ms
100000 promises: 8287.715ms
node(16363,0x7fffe04c03c0) malloc: *** error for object 0xc0000000: pointer being freed was not allocated
*** set a breakpoint in malloc_error_break to debug
fish: 'node perf.js' terminated by signal SIGABRT (Abort)
```

I don't know how long it actually took until node crashed because I went to the kitchen to chop some vegetables. I think it was about 15 minutes.

After:
```
10 promises: 1.619ms
100 promises: 5.351ms
1000 promises: 13.490ms
10000 promises: 182.118ms
100000 promises: 1143.396ms
1000000 promises: 15271.319ms
```

Starting at 1000 promises, this yields a real performance boost of about 40ms. At 100,000 promises, we are looking at a 7x faster runtime and at 1 million promises the current implementation crashes node while the new one finishes in about 15 seconds.

The tests still pass but I'm not sure if you'd like me to write any new tests as part of this change.