@ForbesLindesay 2017-06-04
It would be great if you could add in a test case that runs only in latest node on travis for the 1,000,000 case and maybe assert that it completes in under a minute or something so we don't regress too far on this in the future.

Also, can you remove the yarn lock file?