@cpojer 2017-06-04
I removed the lockfile but I don't trust CI systems fast enough to give us accurate data. I think it's reasonable to expect people to run `node perf.js` once in a while – this library isn't changing all that often :)