@lpinca 2017-06-06
There is also https://github.com/invertase/denque which seems faster than double-ended-queue.