@ForbesLindesay 2017-06-06
Happy to accept a pull request if you benchmark it and it's faster.  I'd also be interested to see someone benchmark the super primitive:

```js
function Queue() {
  this._s1 = [];
  this._s2 = [];
}
Queue.prototype.push = function (value) {
  this._s1.push(value);
};
Queue.prototype.shift = function () {
  if (!this._s2.length) {
    while (this._s1.length) {
      this._s2.push(this._s1.pop());
    }
  }
  return this._s2.pop();
};
Queue.prototype.isEmpty = function () {
  return !(this._s1.length || this._s2.length);
};
```

as we only need a tiny part of the functionality of these larger libraries, and for client side use it may be more important to just use an implementation with as little code as possible.