@ForbesLindesay 2017-06-06
```js
    while (this._s1.length) {
      this._s2.push(this._s1.pop());
    }
```

Could alternatively be something like:

```js
var emptyArray = this._s2;
this._s2 = this._s1.reverse();
this._s1 = emptyArray;
```

I don't know which would turn out faster.