@cpojer 2017-06-06
I tried out `denque` and it is marginally faster on a single run but within the variance of where I'm not sure it's a real difference.