@lpinca 2017-06-06
This is what I got:

```
$ node benchmark/thousand.js 
Platform info:
Darwin 15.6.0 x64
Node.JS 8.0.0
V8 5.8.283.41
Intel(R) Core(TM)2 Duo CPU     T9800  @ 2.93GHz × 2

denque x 15,496,183 ops/sec ±0.74% (86 runs sampled)
double-ended-queue x 10,195,262 ops/sec ±0.69% (89 runs sampled)

$ node benchmark/two_million.js 
Platform info:
Darwin 15.6.0 x64
Node.JS 8.0.0
V8 5.8.283.41
Intel(R) Core(TM)2 Duo CPU     T9800  @ 2.93GHz × 2

denque x 13,896,185 ops/sec ±0.75% (83 runs sampled)
double-ended-queue x 9,966,981 ops/sec ±0.88% (89 runs sampled)
```

Of course these are their benchmarks, not real word use cases. What I know is that `ioredis` is now using `denque`. It was previously using `double-ended-queue`.

I honestly don't care, just wanted to let you know :)