@cpojer 2017-06-06
Seems like you have a better benchmark – do you mind submitting a PR both with a benchmark and with the switch to denque? :)