@lpinca 2017-06-09
See #16.

I've also benchmarked the solution proposed by @ForbesLindesay above.

```
$ node benchmark/thousand.js 
Platform info:
Darwin 15.6.0 x64
Node.JS 8.1.0
V8 5.8.283.41
Intel(R) Core(TM)2 Duo CPU     T9800  @ 2.93GHz × 2

denque x 15,559,931 ops/sec ±0.70% (88 runs sampled)
double-ended-queue x 10,097,522 ops/sec ±0.51% (88 runs sampled)
primitive x 6,693,244 ops/sec ±1.83% (84 runs sampled)

$ node benchmark/two_million.js 
Platform info:
Darwin 15.6.0 x64
Node.JS 8.1.0
V8 5.8.283.41
Intel(R) Core(TM)2 Duo CPU     T9800  @ 2.93GHz × 2

denque x 13,201,392 ops/sec ±2.00% (83 runs sampled)
double-ended-queue x 9,335,689 ops/sec ±1.12% (83 runs sampled)
primitive x 6,717,857 ops/sec ±0.78% (87 runs sampled)
```