@lpinca 2017-06-09
The "while" variant seems to perform very well

```
$ node benchmark/thousand.js 
Platform info:
Darwin 15.6.0 x64
Node.JS 8.1.0
V8 5.8.283.41
Intel(R) Core(TM)2 Duo CPU     T9800  @ 2.93GHz × 2

denque x 15,448,256 ops/sec ±0.66% (87 runs sampled)
double-ended-queue x 10,034,347 ops/sec ±0.64% (86 runs sampled)
primitive x 12,875,835 ops/sec ±2.09% (87 runs sampled)

$ node benchmark/two_million.js 
Platform info:
Darwin 15.6.0 x64
Node.JS 8.1.0
V8 5.8.283.41
Intel(R) Core(TM)2 Duo CPU     T9800  @ 2.93GHz × 2

denque x 13,701,640 ops/sec ±0.73% (84 runs sampled)
double-ended-queue x 9,367,211 ops/sec ±1.19% (82 runs sampled)
primitive x 13,247,861 ops/sec ±0.52% (88 runs sampled)
```